Add content and mock illustrations for "Services" page (#27)
The "Services" page now has descriptions and explanations of all my
services listed. The page is split into three sections: DevSecOps and
Platform Engineering and a description on how projects are set up. I
chose to leave in the mock illustrations for now but they need to be
replaced before the website is published.

Closes #20.